cache service

+++ src/services/postgres/CollaborationsService.js
@@ -4,15 +4,16 @@ const { Pool } = require('pg');
 const InvariantError = require('../../exceptions/InvariantError');
 
 class CollaborationsService {
-  constructor() {
+  constructor(cacheService) {
     this._pool = new Pool();
+    this._cacheService = cacheService;
   }
 
   async addCollaboration(playlistId, userId) {
     const id = `collab-${nanoid(16)}`;
 
     const query = {
-      text: 'INSERT INTO collaborations VALUES($1, $2, $3) RETURNING id',
+      text: 'INSERT INTO collaborations VALUES($1, $2, $3) RETURNING id, user_id',
       values: [id, playlistId, userId],
     };
 
@@ -22,12 +23,16 @@ class CollaborationsService {
       throw new InvariantError('Failed to add collaboration');
     }
 
+    // collaborator added, accessible playlist also changed
+    const owner = result.rows[0].user_id;
+    await this._cacheService.delete(`playlists:${owner}`);
+
     return result.rows[0].id;
   }
 
   async deleteCollaboration(playlistId, userId) {
     const query = {
-      text: 'DELETE FROM collaborations WHERE playlist_id = $1 AND user_id = $2 RETURNING id',
+      text: 'DELETE FROM collaborations WHERE playlist_id = $1 AND user_id = $2 RETURNING id, user_id',
       values: [playlistId, userId],
     };
 
@@ -36,6 +41,10 @@ class CollaborationsService {
     if (!result.rowCount) {
       throw new InvariantError('Failed to delete collaborations');
     }
+
+    // collaborator deleted, accessible playlist also changed
+    const owner = result.rows[0].user_id;
+    await this._cacheService.delete(`playlists:${owner}`);
   }
 
   async verifyCollaborator(playlistId, userId) {
