+++ src/services/postgres/PlaylistsService.js
@@ -6,9 +6,10 @@ const InvariantError = require('../../exceptions/InvariantError');
 const NotFoundError = require('../../exceptions/NotFoundError');
 
 class PlaylistsService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
     this._collaborationService = collaborationService;
+    this._cacheService = cacheService;
   }
 
   async addPlaylist({ name, owner }) {
@@ -25,27 +26,42 @@ class PlaylistsService {
       throw new InvariantError('Fail to add playlist');
     }
 
+    // playlists owned by owner is updated, thus the cache becomes obsolete
+    await this._cacheService.delete(`playlists:${owner}`);
+    console.log(`deleted: playlists:${owner}`);
     return result.rows[0].id;
   }
 
   async getPlaylists(owner) {
-    const query = {
-      text:
-      `SELECT pl.id, pl.name, users.username 
-      FROM playlists as pl 
-      INNER JOIN 
-      users ON pl.owner = users.id
-      LEFT JOIN collaborations on pl.id = collaborations.playlist_id
-      WHERE pl.owner = $1 OR collaborations.user_id = $1`,
-      values: [owner],
-    };
-    const result = await this._pool.query(query);
-    return result.rows;
+    try {
+      // try to find playlists owned by owner in cache
+      const result = await this._cacheService.get(`playlists:${owner}`);
+      return JSON.parse(result);
+    } catch (error) {
+      // not found, fetch from db
+      const query = {
+        text:
+        `SELECT pl.id, pl.name, users.username 
+        FROM playlists as pl 
+        INNER JOIN 
+        users ON pl.owner = users.id
+        LEFT JOIN collaborations on pl.id = collaborations.playlist_id
+        WHERE pl.owner = $1 OR collaborations.user_id = $1`,
+        values: [owner],
+      };
+
+      const result = await this._pool.query(query);
+
+      // update cache w/ data from db
+      await this._cacheService.set(`playlists:${owner}`, JSON.stringify(result.rows));
+      console.log(`added playlists:${owner}`);
+      return result.rows;
+    }
   }
 
   async deletePlaylistById(playlistId) {
     const query = {
-      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id',
+      text: 'DELETE FROM playlists WHERE id = $1 RETURNING id, owner',
       values: [playlistId],
     };
 
@@ -54,6 +70,15 @@ class PlaylistsService {
     if (!result.rows.length) {
       throw new InvariantError('Cannot find playlist by that id');
     };
+
+    // playlist no longer exist, delete playlists owned by owner from cache
+    const owner = result.rows[0].owner;
+    await this._cacheService.delete(`playlists:${owner}`);
+    console.log(`deleted: playlists:${owner}`);
+
+    // playlist no longer exist, songs shouldn't be accessible if someone is trying to fetch it by playlistId
+    await this._cacheService.delete(`songs:${playlistId}`);
+    console.log(`deleted songs:${playlistId} `);
   };
 
   async addSongToPlaylist(playlistId, songId) {
@@ -70,27 +95,42 @@ class PlaylistsService {
       throw new InvariantError('Fail to add playlist');
     }
 
+    // songs inside playlist changed, delete songs cache
+    await this._cacheService.delete(`songs:${playlistId}`);
+    console.log(`deleted songs:${playlistId}`);
     return result.rows[0].id;
   }
 
   async getSongsFromPlaylistByPlaylistId(playlistId) {
-    const query = {
-      text:
-      `SELECT songs.id, songs.title, songs.performer 
-      FROM playlistsongs as pl
-      INNER JOIN songs
-      ON pl.song_id = songs.id
-      WHERE pl.playlist_id = $1`,
-      values: [playlistId],
-    };
+    // try to get songs from playlist in cache
+    try {
+      const result = await this._cacheService.get(`songs:${playlistId}`);
+      return JSON.parse(result);
 
-    const result = await this._pool.query(query);
+    // cannot get from cache, try from db
+    } catch (error) {
+      const query = {
+        text:
+        `SELECT songs.id, songs.title, songs.performer 
+        FROM playlistsongs as pl
+        INNER JOIN songs
+        ON pl.song_id = songs.id
+        WHERE pl.playlist_id = $1`,
+        values: [playlistId],
+      };
 
-    if (!result.rowCount) {
-      throw new NotFoundError('Cannot find playlist by that id');
+      const result = await this._pool.query(query);
+
+      if (!result.rowCount) {
+        throw new NotFoundError('Cannot find playlist by that id');
+      }
+
+      // save to cache
+      this._cacheService.set(`songs:${playlistId}`, JSON.stringify(result.rows));
+      console.log(`added songs:${playlistId}`);
+
+      return result.rows;
     }
-
-    return result.rows;
   }
 
   async deleteSongFromPlaylistBySongId(playlistId, songId) {
@@ -104,7 +144,12 @@ class PlaylistsService {
     if (!result.rowCount) {
       throw new InvariantError('Cannot find song by that id');
     };
+
+    // songs inside playlist changed, delete songs cache
+    await this._cacheService.delete(`songs:${playlistId}`);
+    console.log(`deleted songs:${playlistId}`);
   };
+
   // only accept owner id
   async verifyPlaylistOwner(playlistId, owner) {
     const query = {
